# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.5.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /initiatives:
    get:
      operationId: listInitiatives
      tags: [initiatives]
      summary: List all initiatives
      responses:
        "200":
          description: A list of initiatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Initiative"
    post:
      operationId: createInitiative
      tags: [initiatives]
      summary: Create a new initiative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeInput"
      responses:
        "201":
          description: Created initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
  /initiatives/{id}:
    get:
      operationId: getInitiative
      tags: [initiatives]
      summary: Get a single initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateInitiative
      tags: [initiatives]
      summary: Update an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeUpdate"
      responses:
        "200":
          description: Updated initiative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteInitiative
      tags: [initiatives]
      summary: Delete an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /clients:
    get:
      operationId: listClients
      tags: [execution]
      summary: List clients
      responses:
        "200":
          description: A list of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [execution]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCreate"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [execution]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [execution]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [execution]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /programs:
    get:
      operationId: listPrograms
      tags: [execution]
      summary: List programs
      responses:
        "200":
          description: A list of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [execution]
      summary: Create a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramCreate"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [execution]
      summary: Get a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Not found
    patch:
      operationId: updateProgram
      tags: [execution]
      summary: Update a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Not found
    delete:
      operationId: deleteProgram
      tags: [execution]
      summary: Delete a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /projects:
    get:
      operationId: listProjects
      tags: [execution]
      summary: List projects, newest first
      parameters:
        - name: initiativeId
          in: query
          required: false
          schema:
            type: integer
          description: Filter to projects originating from a given initiative
      responses:
        "200":
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [execution]
      summary: Create a project (client or internal — no client/program required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [execution]
      summary: Get a project with its milestones
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [execution]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [execution]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /projects/{id}/risks:
    get:
      operationId: listProjectRisks
      tags: [governance]
      summary: List risks for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectRisk"
        "404":
          description: Project not found
    post:
      operationId: createProjectRisk
      tags: [governance]
      summary: Add a risk to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRiskCreate"
      responses:
        "201":
          description: Created risk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRisk"
        "404":
          description: Project not found
  /projects/{id}/risks/{riskId}:
    patch:
      operationId: updateProjectRisk
      tags: [governance]
      summary: Update a risk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: riskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRiskUpdate"
      responses:
        "200":
          description: Updated risk
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectRisk"
        "404":
          description: Not found
    delete:
      operationId: deleteProjectRisk
      tags: [governance]
      summary: Delete a risk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: riskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /projects/{id}/approvals:
    get:
      operationId: listProjectApprovals
      tags: [governance]
      summary: List approvals for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectApproval"
        "404":
          description: Project not found
    post:
      operationId: createProjectApproval
      tags: [governance]
      summary: Request an approval on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectApprovalCreate"
      responses:
        "201":
          description: Created approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectApproval"
        "404":
          description: Project not found
  /projects/{id}/approvals/{approvalId}:
    patch:
      operationId: updateProjectApproval
      tags: [governance]
      summary: Update or decide an approval (server stamps decidedAt on decision)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: approvalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectApprovalUpdate"
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectApproval"
        "404":
          description: Not found
    delete:
      operationId: deleteProjectApproval
      tags: [governance]
      summary: Delete an approval request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: approvalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /projects/{id}/readiness:
    get:
      operationId: listReadinessAssessments
      tags: [governance]
      summary: List go-live readiness assessments (with items and computed summary)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReadinessAssessment"
        "404":
          description: Project not found
    post:
      operationId: createReadinessAssessment
      tags: [governance]
      summary: Create a readiness assessment (optionally seeded with standard category items)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadinessAssessmentCreate"
      responses:
        "201":
          description: Created assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessAssessment"
        "404":
          description: Project not found
  /projects/{id}/readiness/{assessmentId}:
    patch:
      operationId: updateReadinessAssessment
      tags: [governance]
      summary: Update a readiness assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: assessmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadinessAssessmentUpdate"
      responses:
        "200":
          description: Updated assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessAssessment"
        "404":
          description: Not found
    delete:
      operationId: deleteReadinessAssessment
      tags: [governance]
      summary: Delete a readiness assessment and its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: assessmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /projects/{id}/readiness/{assessmentId}/items/{itemId}:
    patch:
      operationId: updateReadinessItem
      tags: [governance]
      summary: Update a readiness checklist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: assessmentId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadinessItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessItem"
        "404":
          description: Not found
    delete:
      operationId: deleteReadinessItem
      tags: [governance]
      summary: Delete a readiness checklist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: assessmentId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources with computed capacity/allocation
      responses:
        "200":
          description: Resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceWithCapacity"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a resource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          description: Invalid payload
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Resource detail with current assignments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resource detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDetail"
        "404":
          description: Not found
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Update a resource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource (assignments become department-only demand)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    InitiativeInput:
      type: object
      required:
        - title
        - department
        - submitterName
        - category
        - problemStatement
        - currentProcess
        - desiredOutcome
        - aiConcept
        - prototypeGoal
        - successMetric
      properties:
        title:
          type: string
          minLength: 1
        jiraIssueId:
          type: string
          pattern: '^\d{1,30}$'
          description: "Optional selected Jira issue ID; verified live before initiative creation."
        department:
          type: string
        submitterName:
          type: string
        businessOwner:
          type: string
        executiveSponsor:
          type: string
        category:
          type: string
        status:
          type: string
        executiveSummary:
          type: string
        problemStatement:
          type: string
        currentProcess:
          type: string
        desiredOutcome:
          type: string
        aiConcept:
          type: string
        prototypeGoal:
          type: string
        successMetric:
          type: string
        estimatedHoursSavedMonthly:
          type: number
        estimatedRevenueOpportunity:
          type: number
        estimatedCostSavings:
          type: number
        customerImpact:
          type: string
        complianceRisk:
          type: string
        technicalComplexity:
          type: string
        aiReadiness:
          type: string
        assignedTeam:
          type: string
        currentPhase:
          type: string
        prototypeDay:
          type: integer
...